Move unit testable method to seperate file in market microservice
